+++ src/scripts/scripts_lua.h
@@ -13,9 +13,9 @@
 #include <lua.h>
 #include <lualib.h>
 
-bool script_start(struct t_scripts_state *scripts_state, sds script, struct t_list *arguments,
+bool script_start(struct t_scripts_state *scripts_state, sds scriptname, struct t_list *arguments,
         const char *partition, bool localscript, enum script_start_events start_event,
         unsigned request_id, unsigned long conn_id, sds *error);
-bool script_validate(struct t_config *config, sds name, sds content, sds *error);
+bool script_validate(struct t_config *config, sds scriptname, sds script, sds *error);
 
 #endif